don't jam everything into the caseta module
submodules under caseta (eg caseta::message::*, caseta::connection::*
all make the types a little more obvious)
